+++ qCC/main.cpp
@@ -95,6 +95,14 @@ int main(int argc, char **argv)
 #ifdef CC_GAMEPAD_SUPPORT
 	QGamepadManager::instance(); //potential workaround to bug https://bugreports.qt.io/browse/QTBUG-61553
 #endif
+
+	// Convert the input arguments to QString before the application is initialized
+	// (as it will force utf8, which might prevent from properly reading filenmaes from the command line)
+	QStringList argumentsLocal8Bit;
+	for (int i = 0; i < argc; ++i)
+	{
+		argumentsLocal8Bit << QString::fromLocal8Bit(argv[i]);
+	}
 	
 	ccApplication app(argc, argv, commandLine);
 
@@ -120,9 +128,9 @@ int main(int argc, char **argv)
 	if (commandLine)
 	{
 		//translation file selection
-		if (QString(argv[lastArgumentIndex]).toUpper() == "-LANG")
+		if (argumentsLocal8Bit[lastArgumentIndex].toUpper() == "-LANG")
 		{
-			QString langFilename = QString::fromLocal8Bit(argv[2]);
+			QString langFilename = argumentsLocal8Bit[2];
 
 			ccTranslationManager::Get().loadTranslation(langFilename);
 			commandLine = false;
@@ -165,7 +173,7 @@ int main(int argc, char **argv)
 	if (commandLine)
 	{
 		//command line processing (no GUI)
-		result = ccCommandLineParser::Parse(argc, argv, pluginManager.pluginList());
+		result = ccCommandLineParser::Parse(argumentsLocal8Bit, pluginManager.pluginList());
 	}
 	else
 	{
@@ -198,7 +206,8 @@ int main(int argc, char **argv)
 			QStringList filenames;
 			for (int i = lastArgumentIndex; i < argc; ++i)
 			{
-				QString arg = QString::fromLocal8Bit(argv[i]);
+				QString arg = argumentsLocal8Bit[i];
+
 				//special command: auto start a plugin
 				if (arg.startsWith(":start-plugin:"))
 				{
@@ -227,7 +236,7 @@ int main(int argc, char **argv)
 				}
 				else
 				{
-					filenames << QString::fromLocal8Bit(argv[i]);
+					filenames << arg;
 				}
 			}